cli: split role detection into read and detect
